moved easterEgg to new file "easterEgg"
