+++ src/main.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-use-before-define */
 /// <reference types="jquery" />
 import { hideError } from './error';
+import { easterEgg } from './easterEgg.js';
 
 document.querySelector('#search-pokemon-button').addEventListener('click', () => {
   const pokemon = $('#pokemon-id-input').val().toLowerCase();
@@ -53,6 +54,9 @@ function getPokemonHandler(pokemon) {
   addLoadingToPokemon();
   hideError();
   getPokemonByIdOrName(pokemon);
+  let pokemon;
+  if (pokemonName === 'michelin') {
+    pokemon = easterEgg();
 }
 function addLoadingToPokemon() {
   $('#pokemon-carousel .carousel-inner')
@@ -135,47 +139,6 @@ function getPokemonByIdOrName(pokemonIdOrName) {
       handleError(`The pokemon <strong>${pokemonIdOrName}</strong> does not exist`);
       clearPokemon();
     });
-}
-function easterEgg() {
-  const michelin = {
-    name: 'michelin',
-    types: [
-      {
-        type: {
-          name: 'normal',
-        },
-      },
-    ],
-    abilities: [
-      {
-        ability: {
-          name: 'strong wheels',
-        },
-      },
-    ],
-    moves: [
-      {
-        move: {
-          name: 'change wheel',
-        },
-      },
-      {
-        move: {
-          name: 'wheel slide',
-        },
-      },
-    ],
-    sprites: {
-      other: {
-        michelinWorld: {
-          michelin1: 'img/michelin1.jpeg',
-          michelin2: 'img/michelin2.jpeg',
-          michelin3: 'img/michelin3.jpeg',
-        },
-      },
-    },
-  };
-  pokemonHandler(michelin);
 }
 function pokemonHandler(pokemon) {
   showPokemonInfo();
